+++ src/containers/Navbar/style.scss
@@ -1,0 +1,6 @@
+@import '../../styles/variables';
+nav{
+    a{
+        color: $primaryColor !important;
+    }
+}
